Gui edit (#33)
* Edited the NetworkLauncher scene

Added in the title and the quit button

* Edited the GUI for NetworkLauncher

Change the GUI

* Sound

Added in bgm for the main and map scene, added sound to the buttons in the main scene and map scene and to the shooting of fire. Added the game over scene

* Add the sound

Add the sound file

* Revert "Merge branch 'master' into GUIEdit"

This reverts commit c1b10cbfa5d0187d0695532147ceb78bf5ec015f, reversing
changes made to ffce107ccfea262bd8b21d14fb705847e7be18f0.

* Fix the camera and added in additional physics

Fix the camera and add in some barrels for phyics material. Adjusted the gui a bit

* Added the assets

Added in barrels and creates for physics